requirements.txt
numpy>=1.14.3
pandas>=0.22
tensorflow==2.8
h5py>=2.8
xmljson>=0.1.9
scipy>=1.1
scikit-learn>=0.20
tqdm>=4.26
xarray>=0.11.2
seaborn>=0.9
openpyxl>=3.0
Pillow==10.4.0